+++ src/service/metrics.js
@@ -36,7 +36,9 @@ import {
 } from '~/src/repositories/metrics-repository.js'
 import {
   CalculationTypes,
+  createFormMap,
   dateFallsInsideTimeslot,
+  decodeParamList,
   formatDateOnly,
   getMetricCalcType,
   isDraftSubmission,
@@ -619,19 +621,6 @@ export function decrementCountsForRepublish(map) {
 }
 
 /**
- * Decode param list of encoded strings
- * @param { string[] | undefined } params
- */
-export function decodeParamList(params) {
-  // Decode param list (if applicable)
-  const paramsDecoded = /** @type {string[]} */ ([])
-  if (params) {
-    params.forEach((o) => paramsDecoded.push(decodeURI(o)))
-  }
-  return paramsDecoded.length ? paramsDecoded : undefined
-}
-
-/**
  * Generates a report based on the stored metrics
  * @param {FilterCriteria} filter
  */
@@ -659,17 +648,6 @@ export async function generateReport(filter) {
   } finally {
     await session.endSession()
   }
-}
-
-/**
- * @param {Record<string, number> | undefined} metricValues
- */
-function createFormMap(metricValues) {
-  const formMap = new Map()
-  for (const [formId, count] of Object.entries(metricValues ?? {})) {
-    formMap.set(formId, count)
-  }
-  return formMap
 }
 
 /**
